[Feat] Isolate SearchInputBox component
Why: so we could reuse it
How: Isolate SearchInputBox to different file and use in right places
