+++ src/components/Common/SearchInputBox/SearchInputBox.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { SearchBox, SearchIcon, SearchInput } from "./SearchInputBoxElements";
+
+const SearchInputBox = ({ placeholder, value, onChange }) => {
+    return (
+        <SearchBox>
+            <SearchIcon />
+            <SearchInput type="text" placeholder={placeholder} value={value} onChange={onChange} />
+        </SearchBox>
+    );
+};
+
+export default SearchInputBox;
