+++ src/components/Common/SearchInputBox/index.js
@@ -1,0 +1,2 @@
+import SearchInputBox from "./SearchInputBox";
+export default SearchInputBox;
